Split checks.py (2,881 LOC) into checks_run_relevant.py and checks_lint_fix.py
Decompose the god-module into two cohesive sibling modules:
- checks_run_relevant.py: relevant-checks runner, contains-pins checker, utilities
- checks_lint_fix.py: lint-fix dispatch loop, repair-loop CLI, check-fix loop
- checks.py: thin re-export facade preserving the original public surface

Update complexity-baseline.json and ruff.toml per-file ignores for the new
modules. Update test monkeypatch targets to the correct submodule namespaces.
No behavior change; all wire formats preserved.

Fixes #5774
